# docs/jamma_swagger.yaml
basePath: /api/
definitions:
  api.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
  api.ReplyData-any:
    properties:
      data: {}
      error:
        type: string
    type: object
  api.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
  api.ReplyData-types_Box:
    properties:
      data:
        $ref: '#/definitions/types.Box'
      error:
        type: string
    type: object
  api.ReplyData-types_Game:
    properties:
      data:
        $ref: '#/definitions/types.Game'
      error:
        type: string
    type: object
  api.ReplyData-types_Recharge:
    properties:
      data:
        $ref: '#/definitions/types.Recharge'
      error:
        type: string
    type: object
  api.ReplyData-types_SignIn:
    properties:
      data:
        $ref: '#/definitions/types.SignIn'
      error:
        type: string
    type: object
  api.ReplyData-types_User:
    properties:
      data:
        $ref: '#/definitions/types.User'
      error:
        type: string
    type: object
  api.ReplyList-api_attachInfo:
    properties:
      data:
        items:
          $ref: '#/definitions/api.attachInfo'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_Box:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Box'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_Game:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Game'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_Recharge:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Recharge'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_SignIn:
    properties:
      data:
        items:
          $ref: '#/definitions/types.SignIn'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_User:
    properties:
      data:
        items:
          $ref: '#/definitions/types.User'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.attachInfo:
    properties:
      folder:
        type: boolean
      mime:
        type: string
      name:
        type: string
      size:
        type: integer
      time:
        type: string
    type: object
  api.moveBody:
    properties:
      path:
        type: string
    type: object
  api.renameBody:
    properties:
      name:
        type: string
    type: object
  types.Box:
    properties:
      created:
        type: string
      desc:
        description: 说明
        type: string
      disabled:
        description: 禁用
        type: boolean
      icon:
        description: 图片
        type: string
      id:
        type: integer
      name:
        description: 名称
        type: string
      type:
        description: 类型
        type: string
    type: object
  types.Game:
    properties:
      created:
        type: string
      desc:
        description: 说明
        type: string
      disabled:
        description: 禁用
        type: boolean
      icon:
        description: 图片
        type: string
      id:
        type: integer
      name:
        description: 名称
        type: string
      type:
        description: 类型
        type: string
    type: object
  types.Recharge:
    properties:
      amount:
        description: 充值金额
        type: integer
      created:
        type: string
      id:
        type: integer
      user_id:
        description: 用户ID
        type: integer
    type: object
  types.SignIn:
    properties:
      created:
        type: string
      id:
        type: integer
      user_id:
        description: 用户ID
        type: integer
      username:
        description: 账号
        type: string
    type: object
  types.User:
    properties:
      avatar:
        description: 头像
        type: string
      balance:
        description: 余额
        type: number
      cellphone:
        description: 手机
        type: string
      created:
        type: string
      disabled:
        description: 禁用
        type: boolean
      email:
        description: 邮箱
        type: string
      id:
        type: integer
      integral:
        description: 游戏积分
        type: integer
      name:
        description: 名称
        type: string
      username:
        description: 账号
        type: string
    type: object
info:
  contact: {}
  description: API文档
  title: jamma接口文档
  version: 1.0 版本
paths:
  /attach/list/{name}:
    get:
      description: 查询附件
      parameters:
      - description: 路径
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-api_attachInfo'
      summary: 查询附件
      tags:
      - attach
  /attach/mkdir/{name}:
    get:
      description: 创建目录
      parameters:
      - description: 路径
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-any'
      summary: 创建目录
      tags:
      - attach
  /attach/move/{name}:
    post:
      consumes:
      - application/json
      description: 移动附件
      parameters:
      - description: 路径
        in: path
        name: name
        required: true
        type: string
      - description: 新路径
        in: body
        name: b
        required: true
        schema:
          $ref: '#/definitions/api.moveBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-any'
      summary: 移动附件
      tags:
      - attach
  /attach/remove/{name}:
    get:
      description: 删除附件
      parameters:
      - description: 路径
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-any'
      summary: 删除附件
      tags:
      - attach
  /attach/rename/{name}:
    post:
      consumes:
      - application/json
      description: 重命名附件
      parameters:
      - description: 路径
        in: path
        name: name
        required: true
        type: string
      - description: 新名称
        in: body
        name: b
        required: true
        schema:
          $ref: '#/definitions/api.renameBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-any'
      summary: 重命名附件
      tags:
      - attach
  /attach/upload/{name}:
    post:
      consumes:
      - multipart/form-data
      description: 上传附件（支持多文件，不用特定为file）
      parameters:
      - description: 附件
        in: formData
        name: file
        required: true
        type: file
      - description: 路径
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-any'
      summary: 上传附件
      tags:
      - attach
  /backup/export:
    get:
      consumes:
      - application/json
      description: 导出所有数据
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 导出所有数据
      tags:
      - backup
  /backup/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入所有数据，文件内容为SQL文本，
      parameters:
      - description: SQL
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 导入所有数据
      tags:
      - backup
  /box/{id}:
    get:
      consumes:
      - application/json
      description: 获取游戏机
      parameters:
      - description: 游戏机ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Box'
      summary: 获取游戏机
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 修改游戏机
      parameters:
      - description: 游戏机ID
        in: path
        name: id
        required: true
        type: integer
      - description: 游戏机信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/types.Box'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Box'
      summary: 修改游戏机
      tags:
      - device
  /box/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除游戏机
      parameters:
      - description: 游戏机ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Box'
      summary: 删除游戏机
      tags:
      - device
  /box/count:
    post:
      consumes:
      - application/json
      description: 查询游戏机数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询游戏机数量
      tags:
      - device
  /box/create:
    post:
      consumes:
      - application/json
      description: 创建游戏机
      parameters:
      - description: 游戏机信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Box'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Box'
      summary: 创建游戏机
      tags:
      - device
  /box/list:
    get:
      consumes:
      - application/json
      description: 查询游戏机
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Box'
      summary: 查询游戏机
      tags:
      - device
  /box/search:
    post:
      consumes:
      - application/json
      description: 查询游戏机
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Box'
      summary: 查询游戏机
      tags:
      - device
  /hall/{id}:
    get:
      consumes:
      - application/json
      description: 获取游戏厅
      parameters:
      - description: 游戏厅ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Game'
      summary: 获取游戏厅
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 修改游戏厅
      parameters:
      - description: 游戏厅ID
        in: path
        name: id
        required: true
        type: integer
      - description: 游戏厅信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/types.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Game'
      summary: 修改游戏厅
      tags:
      - device
  /hall/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除游戏厅
      parameters:
      - description: 游戏厅ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Game'
      summary: 删除游戏厅
      tags:
      - device
  /hall/count:
    post:
      consumes:
      - application/json
      description: 查询游戏厅数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询游戏厅数量
      tags:
      - device
  /hall/create:
    post:
      consumes:
      - application/json
      description: 创建游戏厅
      parameters:
      - description: 游戏厅信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Game'
      summary: 创建游戏厅
      tags:
      - device
  /hall/list:
    get:
      consumes:
      - application/json
      description: 查询游戏厅
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Game'
      summary: 查询游戏厅
      tags:
      - device
  /hall/search:
    post:
      consumes:
      - application/json
      description: 查询游戏厅
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Game'
      summary: 查询游戏厅
      tags:
      - device
  /img/upload:
    post:
      consumes:
      - application/json
      description: 可以上传多个图片
      parameters:
      - description: 图片
        in: body
        name: img
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-any'
      summary: 上传图片
      tags:
      - project
  /recharge/count:
    post:
      consumes:
      - application/json
      description: 查询重置记录数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询重置记录数量
      tags:
      - device
  /recharge/create:
    post:
      consumes:
      - application/json
      description: 创建重置记录
      parameters:
      - description: 重置记录信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Recharge'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Recharge'
      summary: 创建重置记录
      tags:
      - device
  /recharge/list:
    get:
      consumes:
      - application/json
      description: 查询重置记录
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Recharge'
      summary: 查询重置记录
      tags:
      - device
  /recharge/search:
    post:
      consumes:
      - application/json
      description: 查询重置记录
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Recharge'
      summary: 查询重置记录
      tags:
      - device
  /user/{id}:
    get:
      consumes:
      - application/json
      description: 获取用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 获取用户
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 修改用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 修改用户
      tags:
      - user
  /user/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 删除用户
      tags:
      - user
  /user/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 禁用用户
      tags:
      - user
  /user/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 启用用户
      tags:
      - user
  /user/count:
    post:
      consumes:
      - application/json
      description: 查询用户数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询用户数量
      tags:
      - user
  /user/create:
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 用户信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 创建用户
      tags:
      - user
  /user/list:
    get:
      consumes:
      - application/json
      description: 查询用户
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_User'
      summary: 查询用户
      tags:
      - user
  /user/me:
    get:
      description: 这里写描述 get users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 获取用户信息
      tags:
      - user
  /user/password:
    get:
      consumes:
      - application/json
      description: 修改密码
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_User'
      summary: 修改密码
      tags:
      - user
  /user/search:
    post:
      consumes:
      - application/json
      description: 查询用户
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_User'
      summary: 查询用户
      tags:
      - user
swagger: "2.0"